requirements.txt
numpy
llvmlite
numba
torch>=1.1
tensorboardX
easydict
pyyaml<5.4
scikit-image
tqdm
torchvision
SharedArray
